Refactoring Span/ReadOnlySpan Extensions into partial classes - Separating shared and platform specific implementations
